Illustration of problem
There is a solution to the game, but probing and average strategy sampling don't reach it. (if we change the random seed in random generator from 10 to 11, then we do.) The problem is that after the first iteration, it looks like player 1 should play 2. Thus, when optimizing player 2, player 1 plays 2. Meanwhile, when optimizing player 1, if player 1 considers 1, then player 2 plays 2. We never have a situation where we optimize player 1 and player 2 tries something else.