starting approach of moving inverted chance decisions to end of game, but too complicated
Problem: the Game determines when the game is complete by using GameProgress, but then the GameDefinition also assesses game completeness without using GameProgress. The GameDefinition SkipDecision function thus must know to skip certain decisions based on GameHistory, and so we've changed GameHistory to allow for provisional completeness. But this is complex and may result in errors.